# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoMo-Nexus API",
    "description": "\n# MoMo-Nexus API\n\nCentral Communication Hub for the MoMo Red Team Ecosystem.\n\n## Overview\n\nNexus serves as the central hub connecting all MoMo ecosystem devices:\n\n- **MoMo** - Field wardriving device (Pi 5)\n- **GhostBridge** - Network implant (NanoPi R2S)\n- **Mimic** - USB HID attack platform (Pi Zero 2W)\n- **Swarm** - LoRa mesh network (Meshtastic)\n\n## Authentication\n\nAll API endpoints require Bearer token authentication:\n\n```\nAuthorization: Bearer <api_key>\n```\n\n## Communication Channels\n\nNexus supports multiple communication channels with automatic failover:\n\n| Channel | Priority | Use Case |\n|---------|----------|----------|\n| LoRa | 1 | Offline, long range |\n| 4G/LTE | 2 | High bandwidth |\n| WiFi | 3 | Local network |\n| BLE | 4 | Short range, low power |\n\n## Real-time Updates\n\nWebSocket endpoint available at `/ws` for real-time event streaming.\n\n## Rate Limits\n\n- API: 100 requests/minute per device\n- WebSocket: 10 messages/second\n- File uploads: 10MB max\n\n---\n\n*MoMo-Nexus - Secure. Stealthy. Synchronized.*\n",
    "contact": {
      "name": "MoMo Team",
      "url": "https://github.com/momo-team"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/status": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Status",
        "description": "Get system status.\n\nReturns current system status including:\n- Online/offline devices count\n- Channel availability\n- Active alerts\n- System uptime",
        "operationId": "get_status_api_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ]
      }
    },
...
    "/api/stats": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Stats",
        "description": "Get detailed system statistics.\n\nReturns aggregated stats including:\n- Total devices registered\n- Active connections per channel\n- Message throughput\n- Capture counts",
        "operationId": "get_stats_api_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ]
      }
    },
    "/api/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "List Devices",
        "description": "List all registered devices.\n\nReturns a list of all devices registered with Nexus.\nSupports filtering by status and device type.\n\nExamples:\n- GET /devices?status=online - All online devices\n- GET /devices?type=momo - All MoMo field units",
        "operationId": "list_devices_api_devices_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (online, offline, unknown)",
              "title": "Status"
            },
            "description": "Filter by status (online, offline, unknown)"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by device type (momo, ghostbridge, mimic, relay)",
              "title": "Type"
            },
            "description": "Filter by device type (momo, ghostbridge, mimic, relay)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "title": "Response List Devices Api Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get Device",
        "description": "Get device details.\n\nReturns detailed information about a specific device including:\n- Current status and last seen time\n- Battery level and version\n- Available communication channels\n- GPS location (if available)",
        "operationId": "get_device_api_devices__device_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Unregister Device",
        "description": "Unregister a device.\n\nRemoves the device from Nexus fleet management.\nThe device will need to re-register to reconnect.",
        "operationId": "unregister_device_api_devices__device_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices/{device_id}/command": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send Command",
        "description": "Send command to a device.\n\nAvailable commands depend on device type:\n\n**MoMo commands:**\n- `scan` - Start WiFi scanning\n- `capture` - Begin handshake capture\n- `stop` - Stop current operation\n- `reboot` - Restart device\n\n**GhostBridge commands:**\n- `beacon` - Send beacon\n- `stealth` - Toggle stealth mode\n\n**Mimic commands:**\n- `execute` - Run payload\n- `switch_mode` - Change USB mode",
        "operationId": "send_command_api_devices__device_id__command_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "cmd",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Command name (scan, capture, stop, reboot, etc.)",
              "title": "Cmd"
            },
            "description": "Command name (scan, capture, stop, reboot, etc.)"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON params",
              "title": "Params"
            },
            "description": "JSON params"
          },
          {
            "name": "wait",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Wait for result",
              "default": true,
              "title": "Wait"
            },
            "description": "Wait for result"
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Timeout seconds",
              "default": 30,
              "title": "Timeout"
            },
            "description": "Timeout seconds"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/command": {
      "post": {
        "summary": "Send Command Body",
        "description": "Send command to a device (POST body).",
        "operationId": "send_command_body_api_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ]
      }
    },
    "/api/broadcast": {
      "post": {
        "summary": "Broadcast Command",
        "description": "Broadcast command to multiple devices.",
        "operationId": "broadcast_command_api_broadcast_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "cmd",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Command name",
              "title": "Cmd"
            },
            "description": "Command name"
          },
          {
            "name": "device_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by device type",
              "title": "Device Type"
            },
            "description": "Filter by device type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages": {
      "post": {
        "summary": "Send Message",
        "description": "Send a message to a device.",
        "operationId": "send_message_api_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ]
      }
    },
...
    "/api/alerts/{alert_id}": {
      "get": {
        "summary": "Get Alert",
        "description": "Get alert details.",
        "operationId": "get_alert_api_alerts__alert_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels": {
      "get": {
        "summary": "List Channels",
        "description": "List all channels and their status.",
        "operationId": "list_channels_api_channels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Enter your API key"
      },
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key in header"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
...
    {
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyHeader": []
    }
  ]
}